First project files and endpoints added

+++ src/main/java/pl/finances/finances_app/repositories/UserRepository.java
@@ -7,11 +7,11 @@ import org.springframework.stereotype.Repository;
 import pl.finances.finances_app.repositories.entities.UserEntity;
 
 @Repository
-public interface UserRepository extends JpaRepository<UserEntity, Integer> {
+public interface UserRepository extends JpaRepository<UserEntity, Long> {
     UserEntity save(UserEntity user);
-    boolean existsById(Integer id);
-    Optional<UserEntity> findById(Integer id);
+    boolean existsById(Long id);
+    Optional<UserEntity> findById(Long id);
     Optional<UserEntity> findByUsername(String username);
-    void deleteById(Integer id);
+    void deleteById(Long id);
     boolean existsByUsername(String username);
 }
